Implement E2E test for 'Report a problem' scenario in all apps. Fix breaking help menu e2e test (because of name unifying on 'Report a problem' dialog).
refs: MBL-18946
affects: Student, Teacher, Parent
release note:
