+++ apps/teacher/src/main/res/values/strings.xml
@@ -711,7 +711,7 @@
     <string name="termsOfUse">Terms of Use</string>
     <string name="searchGuides">Search the Canvas Guides</string>
     <string name="searchGuidesDetails">Find answers to common questions</string>
-    <string name="reportProblem">Report a Problem</string>
+    <string name="reportProblem">Report a problem</string>
     <string name="reportProblemDetails">If the app misbehaves, let us know</string>
     <string name="requestFeature">Request a Feature</string>
     <string name="requestFeatureDetails">Have an idea to improve the app?</string>
